workshop-genai/financial-documents/pdf-build/requirements.txt
python-dotenv
neo4j
neo4j-viz
neo4j-graphrag[openai]
jupyter
langgraph
langchain
langchain_neo4j
langchain_openai
ollama